Split out `generateGalleryShortcode`
